add analytics:fetch-job-latest command, encapsulate reused code in support
