+++ src/commands/analytics/fetch-job-list.ts
@@ -1,20 +1,12 @@
 import ApiCommand from '../../api-command';
-import * as querystring from 'querystring';
-import { flags } from '@oclif/command';
+import * as fetchJobCommand from '../../support/fetch-job-commands';
 
 export default class FetchJobList extends ApiCommand {
     static description = `Return fetch jobs submitted by the user.`;
 
     static flags = {
         ...ApiCommand.flags,
-        'schedule-config-id': flags.integer({
-            description:
-                'A filter used to select jobs created by the given schedule config identfier.',
-        }),
-        'client-context': flags.string({
-            description:
-                'A filter used to select jobs created with the given client context.',
-        }),
+        ...fetchJobCommand.filterFlags,
     };
 
     static examples = [
@@ -25,22 +17,9 @@ export default class FetchJobList extends ApiCommand {
     ];
 
     async run() {
+        const endpoint = `/v20200626/analytics/fetch_job`;
         const opts = this.parse(FetchJobList);
-
-        const queryArgs: { [index: string]: any } = {};
-
-        if (opts.flags['schedule-config-id']) {
-            queryArgs.schedule_config_id = Number(
-                opts.flags['schedule-config-id']
-            );
-        }
-
-        if (opts.flags['client-context']) {
-            queryArgs.client_context = opts.flags['client-context'];
-        }
-
-        const queryStr = querystring.stringify(queryArgs);
-        const path = `/v20200626/analytics/fetch_job?${queryStr}`;
+        const path = fetchJobCommand.processFilters(endpoint, opts);
 
         const res = await this.fetch(path, undefined, `fetching`);
         let cols = {};
